Update workspace settings and cleaned code in index and About sections
This commit implements a few changes. Firstly, it updates the workspace.xml including the number of local tasks. The task 'new skills' is added to show the new skills acquired. Also, a few changes were made to the 'index.tsx' file for better code readability and change in the resume's file name.
In the 'About.tsx' file, languages have been adjusted to reflect current proficiency.